UART y ADC Bien, encoder mal.
UART y ADC andan perfecto, refresca al toque y esta todo re god. No pude hacer que el codigo me capte el encoder. Medi con el adc y el encoder envia bien la senal por el cable pero nunca termina de llegar. Creo que tiene que ver conque usamos los puertos 2.6 y 2.7 que no son tan comunes y a lo mejor esos no estan funcionando bien como input.

+++ ADC.h
@@ -9,7 +9,7 @@
 typedef struct 
 {
     volatile uint16_t value;
-    volatile float voltage;
+    volatile uint16_t voltage;
 }adc_t;
